+++ lib/polyphony.rb
@@ -70,7 +70,7 @@ end
 def install_terminating_signal_handler(signal, exception_class)
   trap(signal) do
     exception = exception_class.new
-    Thread.current.break_out_of_ev_loop(exception)
+    Thread.current.break_out_of_ev_loop(Thread.current.main_fiber, exception)
   end
 end