+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+## 0.32 2020-02-29
+
 * Accept optional throttling rate in `#spin_loop`
 * Remove CancelScope
 * Allow spinning fibers from a parent fiber other than the current